src/function-app/requirements.txt
azure-functions>=1.18.0
azure-servicebus>=7.11.0
